Register a callback to update the Camera's projection matrix when the viewport changes size

+++ Molecules/src/Rendering/InputLayout.h
@@ -13,7 +13,7 @@ public:
 		GFX_THROW_INFO(
 			device->CreateInputLayout(
 				inputElements.data(),
-				inputElements.size(),
+				static_cast<UINT>(inputElements.size()),
 				vs->GetBufferPointer(),
 				vs->GetBufferSize(),
 				&m_inputLayout
